+++ api/src/main/kotlin/org/monsing/record/request/RequestFeedbackRequest.kt
@@ -1,4 +1,4 @@
-package org.monsing.record
+package org.monsing.record.request
 
 data class RequestFeedbackRequest(
     val teacherId: Long
