+++ api/src/main/kotlin/org/monsing/record/request/UploadRecordRequest.kt
@@ -1,4 +1,4 @@
-package org.monsing.record
+package org.monsing.record.request
 
 data class UploadRecordRequest(
     val title: String
