+++ api/src/main/kotlin/org/monsing/record/response/FeedbackResponse.kt
@@ -1,4 +1,4 @@
-package org.monsing.record
+package org.monsing.record.response
 
 import java.time.LocalDateTime
