+++ api/src/main/kotlin/org/monsing/record/response/RecordUploadResponse.kt
@@ -1,4 +1,4 @@
-package org.monsing.record
+package org.monsing.record.response
 
 data class RecordUploadResponse(
     val key: String
